Hack up a ClientConnection and ConnectionManager
Be able to accept incoming connections. The inheritance structure for SMPPConnection and SMPPClientConnection is a bit awkward. The automatic re-connect behavior should be added later (or through traits) instead of adding code to disable it.

The SMPPClientConnection will close the socket after five seconds due the lack of authentication/bind.