Return an error with an empty message id
When using the errorResponseBody selector and the default empty bytearray we would end in a situation that a proper SMPPSubmitSMResponse can't be parsed by the SMPP Codec support anymore. The messageId might need to be something else as well. I didn't read the specification when working on the HomeRouting unit test.